Changes according to comments
