+++ src/App.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import {
-  BrowserRouter,
   Route,
   Switch,
   NavLink,
@@ -11,39 +10,49 @@ import './App.css';
 import HomePage from './components/HomePage/HomePage';
 import TabsPage from './components/TabsPage/TabsPage';
 
+const tabs = [
+  { id: 'tab-1', title: 'Tab 1', content: 'Some text 1' },
+  { id: 'tab-2', title: 'Tab 2', content: 'Some text 2' },
+  { id: 'tab-3', title: 'Tab 3', content: 'Some text 3' },
+];
+
 const App = () => (
-  <BrowserRouter basename="/react_tabs-with-router">
-    <div className="App">
-      <nav>
-        <ul className="main-nav">
-          <li className="main-nav__item">
-            <NavLink
-              className="main-nav__link" 
-              activeClassName="main-nav__link--active" 
-              to="/" 
-              exact
-            >
-              HomePage
+  <div className="App">
+    <nav>
+      <ul className="main-nav">
+        <li className="main-nav__item">
+          <NavLink
+            className="main-nav__link"
+            activeClassName="main-nav__link--active"
+            to="/"
+            exact
+          >
+            HomePage
             </NavLink>
-          </li>
-          <li className="main-nav__item">
-            <NavLink 
-              className="main-nav__link" 
-              activeClassName="main-nav__link--active" 
-              to="/tabs" 
-              exact
-            >
-              TabsPage
+        </li>
+        <li className="main-nav__item">
+          <NavLink
+            className="main-nav__link"
+            activeClassName="main-nav__link--active"
+            to="/tabs"
+            exact
+          >
+            TabsPage
             </NavLink>
-          </li>
-        </ul>
-      </nav>
-      <Switch>
-        <Route path="/" exact component={HomePage} />
-        <Route path="/tabs/:tabId?" exact component={TabsPage} />
-      </Switch>
-    </div>
-  </BrowserRouter>
+        </li>
+      </ul>
+    </nav>
+    <Switch>
+      <Route path="/" exact component={HomePage} />
+      <Route
+        path="/tabs/:tabId?" 
+        exact 
+        render={({ match }) => (
+          <TabsPage match={match} tabs={tabs} />
+        )}
+      />
+    </Switch>
+  </div>
 );
 
 export default App;
