+++ src/components/Tab/Tab.css
@@ -1,0 +1,4 @@
+.tabs__content {
+  border-top: 1px solid rgba(0, 0, 0, 0.1);
+  padding: 20px 0;
+}
