+++ src/index.js
@@ -1,5 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render(
+  <BrowserRouter basename="/react_tabs-with-router">
+    <App />
+  </BrowserRouter>, 
+  document.getElementById('root'));
